+++ base/shared/src/main/scala/scalaz/debug/debug.scala
@@ -1,8 +1,37 @@
 package scalaz
 package debug
 
+import scalaz.ct.ContravariantClass
+
 import language.experimental.macros
 import language.implicitConversions
+
+/** A typeclass describing types which can be meaningfully represented as a `String`.
+ */
+trait DebugClass[A] {
+
+  /** Produce a `String` representation of `a`.
+   */
+  /* todo: when we have a `Cord`, change this to return that instead,
+   * todo: and add a `def debugs(a: A): String = debug(a).toString`
+   */
+  def debug(a: A): String
+}
+
+object DebugClass {
+
+  /** A factory for `Debug` instances which delegates to Scala's `.toString` method.
+   */
+  def fromToString[A]: Debug[A] = instanceOf[DebugClass[A]](_.toString)
+}
+
+trait DebugInstances {
+  implicit final def contravariantDebug: Contravariant[DebugClass] =
+    instanceOf(new ContravariantClass[DebugClass] {
+      def contramap[A, B](r: DebugClass[A])(f: B => A): DebugClass[B] =
+        b => r.debug(f(b))
+    })
+}
 
 trait DebugSyntax {
   implicit final class ToDebugOps[A](self: A) {
