+++ base/shared/src/main/scala/scalaz/std/boolean.scala
