+++ base/shared/src/main/scala/scalaz/std/byte.scala
