+++ base/shared/src/main/scala/scalaz/std/double.scala
