+++ base/shared/src/main/scala/scalaz/std/float.scala
