+++ base/shared/src/main/scala/scalaz/std/short.scala
