+++ base/shared/src/main/scala/scalaz/std/string.scala
