+++ base/shared/src/main/scala/scalaz/std/unit.scala
